refactor: library screen and some other things
